Higher precision conversion table

+++ 02-vars-and-math.c
@@ -4,7 +4,7 @@
 	for fahr = 0, 20, ..., 300 */
 int main()
 {
-	int fahr, celsius;
+	float fahr, celsius;
 	int lower, upper, step;
 
 	lower = 0;
@@ -14,7 +14,7 @@ int main()
 	fahr = lower;
 	while (fahr <= upper) {
 		celsius = 5 * (fahr-32) / 9;
-		printf("%d\t%d\n", fahr, celsius);
+		printf("%g\t%g\n", fahr, celsius);
 		fahr = fahr + step;
 	}
 }
